aider: Fix the error in the `do_chat` method by passing a list of message objects to the `chat` method in the `LLM` class.
# Aider chat conversation:

USER: Hello, welcome to Langchain!
The default language model is gpt-3.5-turbo.
You can change the language model using the 'select_model' command.
You can start chatting with the language model using the 'chat' command.
Type 'help' for more information on commands.
Traceback (most recent call last):
  File "/Users/jonbennett/Dropbox/Coding/langapp/main.py", line 62, in <module>
    LangchainCLI().cmdloop()
  File "/Users/jonbennett/Dropbox/Coding/langapp/main.py", line 32, in __init__
    self.do_chat('')
  File "/Users/jonbennett/Dropbox/Coding/langapp/main.py", line 55, in do_chat
    response = self.current_llm.chat([args])
  File "/Users/jonbennett/Dropbox/Coding/langapp/chains/llm.py", line 12, in chat
    response = openai.ChatCompletion.create(
  File "/Users/jonbennett/miniconda3/lib/python3.10/site-packages/openai/api_resources/chat_completion.py",
 line 25, in create                                                                                            return super().create(*args, **kwargs)
  File "/Users/jonbennett/miniconda3/lib/python3.10/site-packages/openai/api_resources/abstract/engine_api_
resource.py", line 153, in create                                                                              response, _, api_key = requestor.request(
  File "/Users/jonbennett/miniconda3/lib/python3.10/site-packages/openai/api_requestor.py", line 230, in re
quest                                                                                                          resp, got_stream = self._interpret_response(result, stream)
  File "/Users/jonbennett/miniconda3/lib/python3.10/site-packages/openai/api_requestor.py", line 624, in _i
nterpret_response                                                                                              self._interpret_response_line(
  File "/Users/jonbennett/miniconda3/lib/python3.10/site-packages/openai/api_requestor.py", line 687, in _i
nterpret_response_line                                                                                         raise self.handle_error_response(
openai.error.InvalidRequestError: '' is not of type 'object' - 'messages.1'

ASSISTANT: The error message indicates that the `args` passed to the `chat` method in the `LLM` class is not of the expected type. The `chat` method expects a list of message objects, but it seems like an empty string is being passed instead.

To fix this, we need to ensure that the `do_chat` method in the `LangchainCLI` class is passing a list of message objects to the `chat` method of the `LLM` class.

Here is the change that needs to be made in the `main.py` file:

```python
main.py
<<<<<<< ORIGINAL
    def do_chat(self, args):
        response = self.current_llm.chat([args])
        print(response)
=======
    def do_chat(self, args):
        if args:
            response = self.current_llm.chat([{"role": "user", "content": args}])
        else:
            response = self.current_llm.chat([{"role": "system", "content": "You are a helpful assistant."}])
        print(response)
>>>>>>> UPDATED
```

This change ensures that if `args` is not empty, it is wrapped in a dictionary with the role set to "user" before being passed to the `chat` method. If `args` is empty, a system message is passed instead.